+++ commands/dns/zone/update.go
@@ -1,0 +1,77 @@
+package zone
+
+import (
+	"context"
+
+	dns "github.com/ionos-cloud/sdk-go-dns"
+
+	"github.com/spf13/cobra"
+
+	"github.com/ionos-cloud/ionosctl/v6/internal/client"
+	"github.com/ionos-cloud/ionosctl/v6/internal/pointer"
+	"github.com/ionos-cloud/ionosctl/v6/pkg/constants"
+	"github.com/ionos-cloud/ionosctl/v6/pkg/core"
+	"github.com/spf13/viper"
+)
+
+func ZonesPutCmd() *core.Command {
+	cmd := core.NewCommand(context.Background(), nil, core.CommandBuilder{
+		Namespace: "dns",
+		Resource:  "zone",
+		Verb:      "update",
+		Aliases:   []string{"u"},
+		ShortDesc: "Partially modify a zone's properties. This command uses a combination of GET and PUT to simulate a PATCH operation",
+		Example:   "ionosctl dns z update --zone ZONE --name newname.com",
+		PreCmdRun: func(c *core.PreCommandConfig) error {
+			if err := core.CheckRequiredFlags(c.Command, c.NS, constants.FlagZone); err != nil {
+				return err
+			}
+
+			return nil
+		},
+		CmdRun: func(c *core.CommandConfig) error {
+			id, err := Resolve(viper.GetString(core.GetFlagName(c.NS, constants.FlagZone)))
+			if err != nil {
+				return err
+			}
+
+			z, _, err := client.Must().DnsClient.ZonesApi.ZonesFindById(context.Background(), id).Execute()
+			if err != nil {
+				return err
+			}
+
+			if fn := core.GetFlagName(c.NS, constants.FlagName); viper.IsSet(fn) {
+				z.Properties.ZoneName = pointer.From(viper.GetString(fn))
+			}
+			if fn := core.GetFlagName(c.NS, constants.FlagDescription); viper.IsSet(fn) {
+				z.Properties.Description = pointer.From(viper.GetString(fn))
+			}
+			if fn := core.GetFlagName(c.NS, constants.FlagEnabled); viper.IsSet(fn) {
+				z.Properties.Enabled = pointer.From(viper.GetBool(fn))
+			}
+
+			zNew, _, err := client.Must().DnsClient.ZonesApi.ZonesPut(context.Background(), id).
+				ZoneEnsure(dns.ZoneEnsure{Properties: z.Properties}).Execute()
+			if err != nil {
+				return err
+			}
+			return c.Printer.Print(getZonePrint(c, zNew))
+		},
+		InitClient: true,
+	})
+
+	cmd.AddStringFlag(constants.FlagZone, constants.FlagZoneShort, "", constants.DescZone, core.RequiredFlagOption())
+	_ = cmd.Command.RegisterFlagCompletionFunc(constants.FlagZone, func(cmd *cobra.Command, args []string, toComplete string) ([]string, cobra.ShellCompDirective) {
+		return ZonesProperty(func(t dns.ZoneRead) string {
+			return *t.Properties.ZoneName
+		}), cobra.ShellCompDirectiveNoFileComp
+	})
+	cmd.AddStringFlag(constants.FlagName, constants.FlagNameShort, "", "The new name of the DNS zone, e.g. foo.com")
+	cmd.AddStringFlag(constants.FlagDescription, "", "", "The new description of the DNS zone")
+	cmd.AddBoolFlag(constants.FlagEnabled, "", true, "Activate or deactivate the DNS zone")
+
+	cmd.Command.SilenceUsage = true
+	cmd.Command.Flags().SortFlags = false
+
+	return cmd
+}